+++ components/ui/CustomFlatList.tsx
@@ -1,6 +1,9 @@
 import { useScrollHeader } from "@/hooks/useScrollHeader";
 import { appTheme } from "@/utils/appTheme";
 import { Ionicons } from "@expo/vector-icons";
+import { FlashList, FlashListProps, FlashListRef } from "@shopify/flash-list";
+
+
 import React, {
   useCallback,
   useEffect,
@@ -10,22 +13,20 @@ import React, {
 } from "react";
 import {
   ActivityIndicator,
-  FlatList,
-  FlatListProps,
   Platform,
   RefreshControl,
   StyleSheet,
   Text,
   ToastAndroid,
   TouchableOpacity,
-  View,
+  View
 } from "react-native";
 import TitleText from "./TitleText";
-
 type Props<T> = {
+
   data: T[];
-  renderItem: FlatListProps<T>["renderItem"];
-  keyExtractor: FlatListProps<T>["keyExtractor"];
+  renderItem: FlashListProps<T>["renderItem"];
+  keyExtractor: FlashListProps<T>["keyExtractor"];
   refreshing: boolean;
   canRefresh: boolean;
   handleRefresh: () => void;
@@ -37,8 +38,9 @@ type Props<T> = {
   subtitle?: string;
   numColumns?: number;
   showScrollTopButton?: boolean;
-  pageSize?: number; // items per batch
+  pageSize?: number;
 };
+
 
 function CustomFlatList<T>({
   data,
@@ -57,7 +59,7 @@ function CustomFlatList<T>({
   showScrollTopButton = true,
   pageSize = 20,
 }: Props<T>) {
-  const flatListRef = useRef<FlatList<T>>(null);
+const flashListRef = useRef<FlashListRef<T>>(null);
   const { handleScroll, showScrollTop, headerVisible } = useScrollHeader();
 
   const [page, setPage] = useState(1);
@@ -91,7 +93,6 @@ function CustomFlatList<T>({
 
   return (
     <>
-      
       {!canRefresh && cooldown ? (
         <TouchableOpacity
           onPress={onCooldownPress}
@@ -108,7 +109,7 @@ function CustomFlatList<T>({
       {showScrollTop && showScrollTopButton && (
         <TouchableOpacity
           onPress={() =>
-            flatListRef.current?.scrollToOffset({ offset: 0, animated: true })
+            flashListRef.current?.scrollToOffset({ offset: 0, animated: true })
           }
           style={styles.scrollTopButton}
           className="bg-primary dark:bg-dark-primary p-4 rounded-full shadow-lg"
@@ -120,20 +121,17 @@ function CustomFlatList<T>({
       )}
 
       {/* List */}
-      <FlatList
-        ref={flatListRef}
+      <FlashList
+        ref={flashListRef}
         data={paginatedData}
         keyExtractor={keyExtractor}
         renderItem={renderItem}
+      
         contentContainerStyle={styles.listContent}
         onScroll={handleScroll}
         scrollEventThrottle={16}
-        progressViewOffset={100}
-        key={numColumns}
         numColumns={numColumns}
-        columnWrapperStyle={
-          numColumns > 1 ? { justifyContent: "space-between" } : undefined
-        }
+ 
         refreshControl={
           <RefreshControl
             refreshing={canRefresh ? refreshing : false}
@@ -179,10 +177,6 @@ function CustomFlatList<T>({
         }
         onEndReached={loadMore}
         onEndReachedThreshold={0.5}
-        initialNumToRender={10}
-        windowSize={5}
-        maxToRenderPerBatch={10}
-        removeClippedSubviews={false}
       />
     </>
   );